requirements.txt
fastapi
uvicorn
requests
jinja2
python-dotenv
schedule
